+++ service-app/src/main/java/by/ita/je/dto/SocialDto.java
@@ -17,5 +17,5 @@ public class SocialDto {
 
     private Double rating;
 
-    private Long followersCount;
+    private Integer followersCount;
 }
