+++ service-app/src/main/java/by/ita/je/models/Social.java
@@ -16,5 +16,5 @@ public class Social {
 
     private Double rating;
 
-    private Long followersCount;
+    private Integer followersCount;
 }
